Rename main app
