+++ config/release.ini
@@ -1,3 +1,4 @@
 [app]
 css_common_file=css/common.bundle.css
-app_template_file=js/app-template.bundle.js
+app_common_file=js/app-common.bundle.js
+app_main_file=js/app-main.bundle.js
