--- src/template/core/templates/base-react.html
@@ -1,12 +1,0 @@
-{% extends "base.html" %}
-{% load resource from core %}
-
-{% block body %}
-    <div id="app"></div>
-
-    <script src="{% resource 'app_common_file' %}" type="text/javascript"></script>
-
-    {% if initial_state is not None %}{{ initial_state|json_script:"initial-state" }}{% endif %}
-
-    {% block app_script %}{% endblock %}
-{% endblock %}
