+++ src/template/settings.py
@@ -29,7 +29,7 @@ config.read(os.path.join(CONFIG_DIR, '%s.ini' % environment))
 RESOURCES = {
     'css_common_file': config.get('app', 'css_common_file', fallback=None),
     'app_common_file': config.get('app', 'app_common_file', fallback=None),
-    'app_template_file': config.get('app', 'app_template_file', fallback=None),
+    'app_main_file': config.get('app', 'app_main_file', fallback=None),
 }
