+++ webpack.config.js
@@ -11,7 +11,7 @@ const paths = {
 module.exports = {
     mode: 'production',
     entry: {
-        template: ['@babel/polyfill', path.join(paths.JS, 'app.jsx')],
+        main: ['@babel/polyfill', path.join(paths.JS, 'app-main.jsx')],
     },
     output: {
         path: paths.DIST,
